correcting url for adminconsole api client. added check to background service to validate a successful deletion in api.

+++ src/AdminConsole/Services/ApplicationDeletionBackgroundService.cs
@@ -1,6 +1,5 @@
 using AdminConsole.Db;
 using AdminConsole.Models;
-using Microsoft.EntityFrameworkCore;
 
 namespace Passwordless.AdminConsole.Services;
 
@@ -49,11 +48,16 @@ public sealed class ApplicationDeletionBackgroundService : BackgroundService
             var applicationIds = await client.ListApplicationsPendingDeletionAsync();
             foreach (var applicationId in applicationIds)
             {
-                await client.DeleteApplicationAsync(applicationId);
-                var application = new Application { Id = applicationId };
-                db.Entry(application).State = EntityState.Deleted;
-                db.Applications.Remove(application);
-                await db.SaveChangesAsync();
+                if (await client.DeleteApplicationAsync(applicationId))
+                {
+                    var application = new Application { Id = applicationId };
+                    db.Applications.Remove(application);
+                    await db.SaveChangesAsync();
+                }
+                else
+                {
+                    _logger.LogError("Failed to delete application: {appId}", applicationId);
+                }
             }
         }
         catch (Exception e)
